Moved timeUtil to new utils folder
