+++ autoscheduler/frontend/src/components/SchedulingPage/SchedulePreview/MiniSchedule/MiniSchedule.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import Meeting from '../../../../types/Meeting';
-import { LAST_HOUR, FIRST_HOUR } from '../../../../timeUtil';
+import { LAST_HOUR, FIRST_HOUR } from '../../../../utils/timeUtil';
 import DayOfWeek from '../../../../types/DayOfWeek';
 import * as parentStyles from '../../Schedule/Schedule.css';
 import * as styles from './MiniSchedule.css';
