+++ autoscheduler/frontend/src/utils/timeUtil.ts
